+++ airbase/airbase.py
@@ -11,13 +11,7 @@ if sys.version_info >= (3, 11):
 else:
     from typing_extensions import assert_never
 
-from .parquet_api import (
-    Dataset,
-    Session,
-    download_metadata,
-    download_parquet,
-    request_info,
-)
+from .parquet_api import Dataset, Session, download, request_info
 from .summary import DB
 
 
@@ -251,7 +245,7 @@ class AirbaseRequest:
         )
         self.session.raise_for_status = raise_for_status
         asyncio.run(
-            download_parquet(
+            download.parquet(
                 self.session, info, dir, overwrite=not skip_existing
             )
         )
@@ -274,4 +268,4 @@ class AirbaseRequest:
         if self.verbose:
             print(f"Writing metadata to {filepath}...", file=sys.stderr)
 
-        asyncio.run(download_metadata(self.session, filepath))
+        asyncio.run(download.metadata(self.session, filepath))
