+++ airbase/parquet_api/__init__.py
@@ -1,11 +1,6 @@
 from .client import Client
 from .dataset import Dataset, ParquetData, request_info
-from .session import (
-    Session,
-    download_metadata,
-    download_parquet,
-    download_summary,
-)
+from .session import Session
 
 __all__ = [
     "Client",
@@ -13,7 +8,4 @@ __all__ = [
     "ParquetData",
     "request_info",
     "Session",
-    "download_metadata",
-    "download_parquet",
-    "download_summary",
 ]
